+++ src/main/java/com/exp/self/wallet/dto/WalletResponse.java
@@ -2,14 +2,15 @@ package com.exp.self.wallet.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 
 public class WalletResponse {
     public UUID walletId;
-    public long balance;
+    public BigDecimal balance;
 
-    public WalletResponse(UUID walletId, long balance) {
+    public WalletResponse(UUID walletId, BigDecimal balance) {
         this.walletId = walletId;
         this.balance = balance;
     }
